requirements.txt
flask==3.0.3
flask-socketio==5.3.6
python-socketio[asyncio_client]==5.11.3
browser-use==0.1.40
langchain-google-genai>=2.0.11,<5
groq>=1.1.2,<2
python-dotenv==1.0.1
websocket-client>=1.8.0
slack-bolt==1.18.1
